+++ frontend/javascripts/router.tsx
@@ -67,7 +67,7 @@ import {
 import ErrorBoundary from "components/error_boundary";
 import { Store } from "oxalis/singletons";
 import VerifyEmailView from "admin/auth/verify_email_view";
-import { MaintenanceBanner } from "maintenance_view";
+import { MaintenanceBanner } from "maintenance_banner";
 
 const { Content } = Layout;
